requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asyncpg==0.29.0
attrs==23.2.0
bcrypt==4.1.2
billiard==4.2.0
black==24.3.0
build==1.2.1
CacheControl==0.14.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crashtest==0.4.1
cryptography==42.0.5
distlib==0.3.8
dnspython==2.6.1
dulwich==0.21.7
email_validator==2.1.1
fastapi==0.110.1
fastapi-cache2==0.2.1
fastjsonschema==2.19.1
filelock==3.13.3
flower==2.0.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
humanize==4.9.0
idna==3.6
iniconfig==2.0.0
installer==0.7.0
itsdangerous==2.1.2
jaraco.classes==3.4.0
jeepney==0.8.0
Jinja2==3.1.3
keyring==24.3.1
kombu==5.3.7
Mako==1.3.2
MarkupSafe==2.1.5
more-itertools==10.2.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
orjson==3.10.0
packaging==24.0
passlib==1.7.4
pathspec==0.12.1
pendulum==3.0.0
pexpect==4.9.0
pkginfo==1.10.0
platformdirs==4.2.0
pluggy==1.4.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
prometheus_client==0.20.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pycparser==2.22
pydantic==2.6.4
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
PyJWT==2.8.0
pyproject_hooks==1.0.0
pytest==8.1.1
pytest-asyncio==0.23.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.8.1
redis==4.6.0
requests==2.31.0
requests-toolbelt==1.0.0
SecretStorage==3.3.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
tomlkit==0.12.4
tornado==6.4
trove-classifiers==2024.3.25
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
vine==5.1.0
virtualenv==20.25.1
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
yarl==1.9.4
